feat: add BackupBackingImage component
