+++ src/routes/backingImage/BackupBackingImage.less
@@ -1,0 +1,3 @@
+.container {
+    padding: 12px;
+}
